requirements.txt
fastapi>=0.111.0
strawberry-graphql[fastapi]>=0.227.0
pydantic>=2.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
pytest>=8.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
